Support Grid and Arc with a beatfile
Grid shows track progress on the top and creates an interactive lightshow on
the rest of the Grid.

Arc 1 shows track progress, Arc 2 is track's filter cutoff, Arc 3 is track's
filter resonance, and Arc 4 is track's volume (-12dB to 0dB).

A beatfile is a file where each line is 1/10th of a second. It can have events
like "B" (bar start), "b" (beat), "k" (kick drum).  Experience shows that
1/10th of a second is not enough for convincing lightshows so this is bound to
change in future iterations.